partial refactor of unused_finder to move parsing-related utilities to a submodule
