+++ crates/unused_finder/src/lib.rs
@@ -15,10 +15,8 @@ extern crate napi_derive;
 extern crate pretty_assertions;
 
 mod core;
-mod export_collector_tests;
 pub mod graph;
-pub mod import_export_info;
-pub mod node_visitor;
+pub mod parse;
 pub mod unused_finder;
 pub mod unused_finder_visitor_runner;
 mod utils;
